requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-mptt==0.16.0